Separated blacklist token feature into subapp
