+++ tests/conftest.py
@@ -37,6 +37,7 @@ def pytest_configure():
 
             'tests',
             'rest_framework_jwt',
+            'rest_framework_jwt.blacklist',
         ),
         PASSWORD_HASHERS=(
             'django.contrib.auth.hashers.MD5PasswordHasher',
