+++ tests/test_authentication.py
@@ -35,9 +35,9 @@ from rest_framework.test import APIRequestFactory, APIClient
 from rest_framework.views import APIView
 
 from rest_framework_jwt import utils
-from rest_framework_jwt.models import JWTBlacklistToken
 from rest_framework_jwt.settings import api_settings, DEFAULTS
 from rest_framework_jwt.authentication import JSONWebTokenAuthentication
+from rest_framework_jwt.blacklist.models import JWTBlacklistToken
 
 try:
     from rest_framework.authentication import oauth2_provider
